remove done state and refactor store
