+++ src/main/index.ts
@@ -6,9 +6,7 @@ import { createWindow } from './utils/createWindow';
 import setupIpcComs from './utils/ipcComms';
 import { buildContextMenu, buildMenuBar } from './utils/menuUtil';
 import initSentry from './utils/sentry';
-import Store from 'electron-store';
 import { PROD_HOST_URL, RENDERER_OUTPUT_DIR } from '../config';
-import { uploadStoreSchema } from './utils/upload';
 
 if (isDev) {
     const electronReload = require('electron-reload');
@@ -40,10 +38,6 @@ export const setIsUpdateAvailable = (value: boolean): void => {
 const serveNextAt = require('next-electron-server');
 serveNextAt(PROD_HOST_URL, {
     outputDir: RENDERER_OUTPUT_DIR,
-});
-const store = new Store({
-    name: 'upload-status',
-    schema: uploadStoreSchema,
 });
 
 // Disable error dialogs by overriding
