+++ src/main/utils/upload.ts
@@ -1,25 +1,8 @@
-import Store, { Schema } from 'electron-store';
 import path from 'path';
 import * as fs from 'promise-fs';
 import { FILE_STREAM_CHUNK_SIZE } from '../../config';
-import { ElectronFile, StoreType } from '../types';
-
-export const uploadStoreSchema: Schema<StoreType> = {
-    filePaths: {
-        type: 'array',
-        items: {
-            type: 'string',
-        },
-    },
-    collectionName: {
-        type: 'string',
-    },
-};
-
-const store = new Store({
-    name: 'upload-status',
-    schema: uploadStoreSchema,
-});
+import { uploadStatusStore } from '../services/store';
+import { ElectronFile } from '../types';
 
 // https://stackoverflow.com/a/63111390
 export const getFilesFromDir = async (dirPath: string) => {
@@ -87,32 +70,23 @@ export async function getElectronFile(filePath: string): Promise<ElectronFile> {
 
 export const setToUploadFiles = (
     filePaths: string[],
-    collectionName: string,
-    done: boolean
+    collectionName: string
 ) => {
-    if (done) {
-        store.delete('filePaths');
-        store.delete('collectionName');
-    } else {
-        store.set('filePaths', filePaths);
-        store.set('collectionName', collectionName);
-    }
+    uploadStatusStore.set('filePaths', filePaths);
+    uploadStatusStore.set('collectionName', collectionName);
 };
 
 export const getPendingUploads = async () => {
-    const filePaths = store.get('filePaths') as string[];
-    const collectionName = store.get('collectionName') as string;
+    const filePaths = uploadStatusStore.get('filePaths') as string[];
+    const collectionName = uploadStatusStore.get('collectionName') as string;
     return {
-        files: await Promise.all(filePaths.map(getElectronFile)),
+        files: filePaths
+            ? await Promise.all(filePaths.map(getElectronFile))
+            : [],
         collectionName,
     };
 };
 
-export const hasPendingUploads = async () => {
-    const pendingFiles = store.get('filePaths');
-    return pendingFiles && pendingFiles.length > 0;
-};
-
 export const updatePendingUploadsFilePaths = (filePaths: string[]) => {
-    store.set('filePaths', filePaths);
+    uploadStatusStore.set('filePaths', filePaths);
 };
